Use a lexicon normal test for static methods
